Починил баг с рекурсией при запросе GetFavoritesForUserAsync

+++ WebAPI/Aplication/DTOs/FavoriteDTO.cs
@@ -3,7 +3,7 @@
     public class FavoriteDTO
     {
         public DateTime FavoritedAt { get; set; }
-        public int PlaceId { get; set; }
+        public ulong PlaceId { get; set; }
         public virtual PlaceDTODefaultCard Place { get; set; } = null!;
     }
 }
