+++ WebAPI/Aplication/DTOs/HistoryDTO.cs
@@ -4,7 +4,7 @@
     {
         public DateTime VisitDateTime { get; set; }
         public bool IsFromRecs { get; set; }
-        public int PlaceId { get; set; }
+        public ulong PlaceId { get; set; }
         public virtual PlaceDTODefaultCard Place { get; set; } = null!;
     }
 }
