+++ WebAPI/Entities/Models/Favorite.cs
@@ -4,15 +4,11 @@ namespace Entities.Models
 {
     public class Favorite : IDbEntity
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
         public DateTime FavoritedAt { get; set; }
 
-
-
-        public int UserId { get; set; }
-        public int PlaceId { get; set; }
-
-        // Связь с User и Place
+        public ulong UserId { get; set; }
+        public ulong PlaceId { get; set; }
 
         public virtual User User { get; set; } = null!;
         public virtual Place Place { get; set; } = null!;
