+++ WebAPI/Infrastructure/Repository/AdHashtagRepository.cs
@@ -6,13 +6,13 @@ namespace Infrastructure.Repository
     public class AdHashtagRepository(MyDbContext _context)
     {
 
-        public async Task AddHashtagToPlaceAd(int placeId, int hashTagId)
+        public async Task AddHashtagToPlaceAd(ulong placeId, ulong hashTagId)
         {
             var record = new AdHashtag() { PlaceId = placeId, HashtagId = hashTagId };
             _context.AdHashtags.AddAsync(record);
             await _context.SaveChangesAsync();
         }
-        public async Task<bool> IsExistAsync(int placeId, int hashTagId)
+        public async Task<bool> IsExistAsync(ulong placeId, ulong hashTagId)
         {
             return await _context.AdHashtags.AnyAsync(a => a.PlaceId == placeId && a.HashtagId == hashTagId);
         }
